feat: middleware test fixture
